python/requirements.txt
grpcio==1.27.1
protobuf==3.11.2
